Move logo implementation to its own class
